requirements.txt
lxml==4.6.3
pydbus==0.6.0
evdev==1.4.0
PyGObject==3.42.0
cryptography==3.3.2
Cython==0.29.24
ipaddr==2.2.0
numpy==1.21.2
ordereddict==1.1
protobuf==3.18.1
pyOpenSSL==21.0.0
pyserial==3.5
evdev==1.4.0
paho_mqtt==1.6.1
pulsectl==21.10.5
